extract BBP implementation to module
